added improvements to individual band sites
so far, only 'jetwad' as a working page. since this is the band i know most about, i decided to fill out as much information about them and then format that content to fit my design concept. i've included a sidebar that holds key information about the group as well as structured the main content of the page.